refactor(planning): extract the mobile planning sheets (#467)
Final mobile planning slice. Move the bottom-sheet dialogs — the shared Sheet
overlay wrapper and the salary / delete-plan / other-expense / simple-override
sheets — into a new `planningSheets.tsx`. The income/delete/other-expense writes
still go through planActions; the override sheet hands its amount up to the parent.
Layout + local form state only. No behavior or markup change; the now-unused
imports (useEffect, useRef, formatIntVN/parseIntVN, TrashIcon, and the planActions
writers) drop out of the view.

Parity: the PlanningClient / MobilePlanningView component tests + the planning E2E
(create plan via SalarySheet, month navigation).

Verified: full unit suite 1261 passed; typecheck 0 errors; lint 0 errors;
git diff --check clean; E2E planning (3) passed. MobilePlanningView 910 → 525 lines
(1564 → 525 across the mobile split).
